+++ obj/src/clients/GrpcClient.js
@@ -195,8 +195,8 @@ class GrpcClient {
                     let protoLoader = require('@grpc/proto-loader');
                     let options = this._packageOptions || {
                         keepCase: true,
-                        // longs: String,
-                        // enums: String,
+                        longs: Number,
+                        enums: Number,
                         defaults: true,
                         oneofs: true
                     };