+++ obj/src/services/index.d.ts
@@ -4,3 +4,4 @@
  */
 export { GrpcEndpoint } from './GrpcEndpoint';
 export { GrpcService } from './GrpcService';
+export { CommandableGrpcService } from './CommandableGrpcService';