+++ obj/src/services/index.js
@@ -8,7 +8,8 @@ var GrpcEndpoint_1 = require("./GrpcEndpoint");
 exports.GrpcEndpoint = GrpcEndpoint_1.GrpcEndpoint;
 var GrpcService_1 = require("./GrpcService");
 exports.GrpcService = GrpcService_1.GrpcService;
-// export { CommandableGrpcService } from './CommandableGrpcService';
+var CommandableGrpcService_1 = require("./CommandableGrpcService");
+exports.CommandableGrpcService = CommandableGrpcService_1.CommandableGrpcService;
 // export { HeartbeatGrpcService } from './HeartbeatGrpcService';
 // export { StatusGrpcService } from './StatusGrpcService';
 //# sourceMappingURL=index.js.map