+++ src/services/index.ts
@@ -4,6 +4,6 @@
  */
 export { GrpcEndpoint } from './GrpcEndpoint';
 export { GrpcService } from './GrpcService';
-// export { CommandableGrpcService } from './CommandableGrpcService';
+export { CommandableGrpcService } from './CommandableGrpcService';
 // export { HeartbeatGrpcService } from './HeartbeatGrpcService';
 // export { StatusGrpcService } from './StatusGrpcService';