+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 ## <a name="3.0.0"></a> 3.0.0 (2019-04-16)
 
 ### Features
-* **clients** GrpcClient
-* **services** GrpcEndpoint, GrpcService
+* **clients** GrpcClient, CommandableGrpcClient
+* **services** GrpcEndpoint, GrpcService, CommandableGrpcService